add invitation system

+++ app/routes/users._index.tsx
@@ -1,11 +1,12 @@
-import { ActionIcon, Badge, Button, Center, Container, Group, Popover, Select, Stack, Text } from "@mantine/core";
+import { ActionIcon, Badge, Button, Center, Container, Group, List, Popover, Select, Stack, Text, TextInput, Title } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
+import { showNotification } from "@mantine/notifications";
 import { ActionFunction, LoaderFunction, json, redirect } from "@remix-run/cloudflare";
-import { Form, useLoaderData, useSubmit } from "@remix-run/react";
-import { IconClick, IconSquareKey, IconUserMinus } from "@tabler/icons-react";
-import { User } from "kysely-codegen";
+import { Form, useActionData, useLoaderData, useSubmit } from "@remix-run/react";
+import { IconClick, IconSquareKey, IconTrash, IconUserMinus, IconX } from "@tabler/icons-react";
+import { User, UserInvite } from "kysely-codegen";
 import { DataTable, DataTableColumn } from "mantine-datatable";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { useErrorNotification } from "~/components/useErrorNotification";
 import { authenticator } from "~/lib/auth.server";
 import { PERMISSION_ERROR, userRoles } from "~/lib/constants";
@@ -19,10 +20,13 @@ export const loader: LoaderFunction = async ({ request, context }) => {
         return json({ users: [], error: PERMISSION_ERROR }, { status: 403 });
     }
     const db = getDB(context);
-    const result = await db.selectFrom('user')
+    const users = await db.selectFrom('user')
         .selectAll()
         .execute();
-    return json({ users: result });
+    const invites = await db.selectFrom('user_invite')
+        .selectAll()
+        .execute();
+    return json({ users: users, invites: invites });
 }
 
 export const action: ActionFunction = async ({ request, context }) => {
@@ -34,40 +38,83 @@ export const action: ActionFunction = async ({ request, context }) => {
     }
     const formData = await request.formData();
     const action = formData.get("action");
-    const userId = formData.get("uid")?.toString();
-    const email = formData.get("email");
-    const role = formData.get("role")?.toString();
-    if (userId) {
-        if (action == "delete" && email != "dserink@gmail.com") {
-            const db = getDB(context);
-            await db.deleteFrom('signin_event')
-                .where('uid', '=', userId)
-                .execute();
-            await db.deleteFrom('user')
-                .where('uid', '=', userId)
-                .execute();
-            console.log(`deleting user with email ${email}`);
+
+    switch (action) {
+        case "delete_user": {
+            const userId = formData.get("uid")?.toString();
+            const email = formData.get("email");
+            if (userId && email != "dserink@gmail.com") {
+                const db = getDB(context);
+                await db.deleteFrom('signin_event')
+                    .where('uid', '=', userId)
+                    .execute();
+                await db.deleteFrom('user')
+                    .where('uid', '=', userId)
+                    .execute();
+                console.log(`deleting user with email ${email}`);
+            }
+            return json({ success: true });
         }
-        else if (action == "set_role") {
-            if (role) {
+        case "set_role": {
+            const userId = formData.get("uid")?.toString();
+            const role = formData.get("role")?.toString();
+            if (userId && role) {
                 const db = getDB(context);
                 await db.updateTable('user')
-                    .set({ role: role})
+                    .set({ role: role })
                     .where('uid', '=', userId)
                     .execute();
                 console.log(`Set role '${role}' on uid '${userId}'`);
             }
+            return json({ success: true });
+        }
+        case "create_invite": {
+            const inviteEmail = formData.get("invite_email")?.toString();
+            const inviteName = formData.get("invite_name")?.toString();
+            const inviteRole = formData.get("invite_role")?.toString();
+            if (inviteEmail && inviteRole) {
+                const db = getDB(context);
+                try {
+                    await db.insertInto('user_invite')
+                        .values({
+                            email: inviteEmail,
+                            display_name: inviteName || null,
+                            role: inviteRole || null,
+                            invited_by_uid: user.uid,
+                            invited_by_name: user.display_name ?? "",
+                            token_expires: new Date(Date.now() + 365 * 24 * 60 * 60 * 1000).toISOString(), // 365 days from now
+                        })
+                        .execute();
+                } catch (error) {
+                    if (error instanceof Error) console.log(error.message);
+                    return json({ success: false, message: `Could not create invite. If this email is already invited, delete it first to re-invite.` }, { status: 500 });
+                }
+                return json({ success: true, message: `Invite created.` });
+            }
+            break;
+        }
+        case "delete_invite": {
+            const inviteEmail = formData.get("inviteId")?.toString();
+            if (inviteEmail) {
+                const db = getDB(context);
+                await db.deleteFrom('user_invite')
+                    .where('email', '=', inviteEmail)
+                    .execute();
+                return json({ success: true, message: `Invite deleted.` });
+            }
+            break;
         }
     }
     return redirect("/users");
 }
 
 export default function UsersIndex() {
-    const data = useLoaderData<{ users: User[], error?: string }>();
+    const { users, invites, error } = useLoaderData<{ users: User[]; invites: UserInvite[]; error?: string }>();
+    const actionData = useActionData<{ success?: boolean; message?: string }>();
     const submit = useSubmit();
     const [isOpen, { close, toggle }] = useDisclosure(false);
     const [selectedRole, setSelectedRole] = useState<string | null>("");
-    useErrorNotification(data.error);
+    useErrorNotification(error);
 
     const handleRoleSave = (uid: string | undefined) => {
         const formData = new FormData();
@@ -78,10 +125,30 @@ export default function UsersIndex() {
         close();
     };
 
+    useEffect(() => {
+        if (actionData && actionData?.success) {
+            showNotification({
+                title: "Success",
+                message: actionData.message,
+                color: "green",
+                autoClose: 3000,
+            });
+        }
+        else if (actionData && actionData?.success == false) {
+            showNotification({
+                title: 'Error',
+                message: actionData.message,
+                color: 'red',
+                icon: <IconX />,
+                autoClose: 3000,
+            });
+        }
+    }, [actionData]);
+
     const renderActions: DataTableColumn['render'] = (record: Partial<User>) => (
         <Group gap={4} wrap="nowrap">
             <Form method="post">
-                <input type="hidden" name="action" value="delete" />
+                <input type="hidden" name="action" value="delete_user" />
                 <input type="hidden" name="uid" value={record.uid} />
                 <input type="hidden" name="email" value={record.email ?? ""} />
                 <ActionIcon
@@ -136,37 +203,88 @@ export default function UsersIndex() {
 
     return (
         <Container size="xl" p="md">
-            <DataTable
-                withTableBorder
-                borderRadius="sm"
-                withColumnBorders
-                striped
-                highlightOnHover
-                records={data.users}
-                columns={[
-                    {
-                        accessor: "display_name",
-                        render: (record) =>
-                            <Group key={record.uid}>
-                                <Text>{record.display_name}</Text>
-                                <Badge size="xs" color="sector-color">{record.role}</Badge>
-                            </Group>,
-                    },
-                    {
-                        accessor: "email",
-                    },
-                    {
-                        accessor: "email_verified",
-                    },
-                    {
-                        accessor: "actions",
-                        title: (<Center><IconClick size={16} /></Center>),
-                        width: '0%',
-                        render: renderActions,
-                    },
-                ]}
+            <Stack>
+                <Title order={2}>Users</Title>
+                <DataTable
+                    withTableBorder
+                    borderRadius="sm"
+                    withColumnBorders
+                    striped
+                    highlightOnHover
+                    records={users}
+                    columns={[
+                        {
+                            accessor: "display_name",
+                            render: (record) =>
+                                <Group key={record.uid}>
+                                    <Text>{record.display_name}</Text>
+                                    <Badge size="xs" color="sector-color">{record.role}</Badge>
+                                </Group>,
+                        },
+                        {
+                            accessor: "email",
+                        },
+                        {
+                            accessor: "email_verified",
+                        },
+                        {
+                            accessor: "actions",
+                            title: (<Center><IconClick size={16} /></Center>),
+                            width: '0%',
+                            render: renderActions,
+                        },
+                    ]}
 
-            />
+                />
+                <Title order={2}>Invite New Users</Title>
+                <Form method="post">
+                    <input type="hidden" name="action" value="create_invite" />
+                    <Stack>
+                        <TextInput
+                            name="invite_email"
+                            label="Email"
+                            required
+                        />
+                        <TextInput
+                            name="invite_name"
+                            label="Name (optional)"
+                        />
+                        <Select
+                            name="invite_role"
+                            label="Role"
+                            data={userRoles}
+                            required
+                        />
+                        <Button type="submit">Create Invite</Button>
+                    </Stack>
+                </Form>
+                <Title order={2}>Pending Invitations</Title>
+                <List spacing="xs">
+                    {invites.map((invite) => (
+                        <List.Item key={invite.email}>
+                            <Group>
+                                <Group>
+                                    <Text><strong>Email:</strong> {invite.email}</Text>
+                                    <Text><strong>Name:</strong> {invite.display_name || 'N/A'}</Text>
+                                    <Badge>{invite.role}</Badge>
+                                    <Text><strong>Invited by:</strong> {invite.invited_by_name}</Text>
+                                    <Text><strong>Invitation Expires:</strong> {new Date(invite.token_expires as string).toLocaleString()}</Text>
+                                </Group>
+                                <Form method="post">
+                                    <input type="hidden" name="action" value="delete_invite" />
+                                    <input type="hidden" name="inviteId" value={invite.email} />
+                                    <ActionIcon
+                                        color="red"
+                                        type="submit"
+                                    >
+                                        <IconTrash size={16} />
+                                    </ActionIcon>
+                                </Form>
+                            </Group>
+                        </List.Item>
+                    ))}
+                </List>
+            </Stack>
         </Container>
     );
 }
